agent/requirements.txt
anthropic>=0.30.0
requests>=2.31.0
python-dotenv>=1.0.0
urllib3>=2.0.0
